% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-cpx-ggplot.R
\name{as_cpx_ggplot}
\alias{as_cpx_ggplot}
\alias{as_cpx_ggplot.ggplot}
\alias{as_cpx_ggplot.new_annotation}
\title{Coerce to a Complex Ggplot}
\usage{
as_cpx_ggplot(plot)

\method{as_cpx_ggplot}{ggplot}(plot)

\method{as_cpx_ggplot}{new_annotation}(plot)
}
\arguments{
\item{plot}{a ggplot object.}
}
\value{
a cpx_ggplot object.
}
\description{
Functions to coerce a object to a cpx_ggplot if possible.
}
\references{
Guangchuang Yu (2020). aplot: Decorate a 'ggplot' with Associated
Information. R package version 0.0.6.https://CRAN.R-project.org/package=aplot
}
