% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-build.R
\name{cpx_ggplot_build}
\alias{cpx_ggplot_build}
\alias{cpx_ggplot_build.cpx_ggplot}
\alias{cpx_ggplot_build.new_annotation}
\title{Build Complex Ggplot}
\usage{
cpx_ggplot_build(plot)

\method{cpx_ggplot_build}{cpx_ggplot}(plot)

\method{cpx_ggplot_build}{new_annotation}(plot)
}
\arguments{
\item{plot}{a cpx_ggplot object.}
}
\value{
a patchwork object.
}
\description{
trans a cpx_ggplot object to patchwork.
}
\references{
Guangchuang Yu (2020). aplot: Decorate a 'ggplot' with Associated
Information. R package version 0.0.6.https://CRAN.R-project.org/package=aplot
}
