% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-funs.R
\name{new_ggplot_r}
\alias{new_ggplot_r}
\alias{new_ggplot_l}
\alias{new_ggplot_b}
\alias{new_ggplot_t}
\title{Helper Functions}
\usage{
new_ggplot_r(...)

new_ggplot_l(...)

new_ggplot_b(...)

new_ggplot_t(...)
}
\arguments{
\item{...}{other parameters.}
}
\value{
a modified gg object.
}
\description{
A wrapper of \code{new_ggplot()} and \code{new_plot()}.
}
