% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-ggplot.R
\name{new_ggplot}
\alias{new_ggplot}
\title{Init a Annotation Plot}
\usage{
new_ggplot(
  data = NULL,
  mapping = aes(),
  width = NULL,
  height = NULL,
  dir = "right",
  inherit.aes = FALSE,
  plot = NULL
)
}
\arguments{
\item{data}{NULL or a data frame.}

\item{mapping}{default list of aesthetic mappings to use for plot.}

\item{width}{a positive numeric, indicating the width of a row annotation.}

\item{height}{a positive numeric, indicating the height of a column annotation.}

\item{dir}{on of "right", "left", "top" or "bottom", indicating the position
of the annotation plot.}

\item{inherit.aes}{if FALSE, overrides the default aesthetics, rather than
combining with them.}

\item{plot}{NULL or a ggplot object.}
}
\value{
a modified gg object.
}
\description{
\code{new_ggplot()} initializes a ggplot object, and all
subsequent elements are part of the annotation diagram.
}
