% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-label.R
\name{new_label}
\alias{new_label}
\title{Add Label Annotation}
\usage{
new_label(
  label,
  width = 0.2,
  height = 0.2,
  dir = "right",
  element_height = NULL,
  element_width = NULL,
  ...
)
}
\arguments{
\item{label}{a character vector.}

\item{width}{a positive numeric, indicating the width of a row annotation.}

\item{height}{a positive numeric, indicating the height of a column annotation.}

\item{dir}{on of "right", "left", "top" or "bottom", indicating the position
of the annotation plot.}

\item{element_height}{the panel height of each label.}

\item{element_width}{the panel width of each label.}

\item{...}{other parameters for \code{grid::textGrob()}.}
}
\value{
a modified gg object.
}
\description{
Functions to add a text label.
}
