% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-plot.R
\name{new_plot}
\alias{new_plot}
\alias{new_plot.ggplot}
\alias{new_plot.grob}
\alias{new_plot.formula}
\alias{new_plot.default}
\title{Init a Annotation Plot}
\usage{
new_plot(plot, ...)

\method{new_plot}{ggplot}(plot, ...)

\method{new_plot}{grob}(plot, ...)

\method{new_plot}{formula}(plot, ...)

\method{new_plot}{default}(plot, ...)
}
\arguments{
\item{plot}{can be a ggplot, grob or base plot object.}

\item{...}{other parameters.}
}
\value{
a modified gg object.
}
\description{
This function, similar to new_ggplot, is used to initialize an
annotated graph.
}
