% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-align.R
\name{scale_align}
\alias{scale_align}
\title{Align the Annotation Plot}
\usage{
scale_align(limits = NULL)
}
\arguments{
\item{limits}{limits of coordinate.}
}
\value{
a gg object.
}
\description{
Align the annotation plot by the main plot.
}
