% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_no_y}
\alias{theme_no_y}
\alias{theme_no_x}
\title{Theme}
\usage{
theme_no_y(...)

theme_no_x(...)
}
\arguments{
\item{...}{other parameters passing to \code{theme()}.}
}
\value{
a theme object.
}
\description{
Theme for removing the x-axis or the y-axis.
}
