% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_no_y}
\alias{theme_no_y}
\alias{theme_no_x}
\title{Theme}
\usage{
theme_no_y(..., plot.margin = margin())

theme_no_x(..., plot.margin = margin())
}
\arguments{
\item{...}{other parameters passing to \code{theme()}.}

\item{plot.margin}{the margin of plot.}
}
\value{
a theme object.
}
\description{
Theme for removing the x-axis or the y-axis.
}
